# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/health": {
      "get": {
        "operationId": "AppController_getHealth",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/profile": {
      "get": {
        "operationId": "AppController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_getMe",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/accounts": {
      "post": {
        "operationId": "AccountsController_create",
        "summary": "Yeni hesap ekle",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAccountDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Hesap başarıyla oluşturuldu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "accounts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "operationId": "AccountsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "accounts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/accounts/{id}": {
      "get": {
        "operationId": "AccountsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "accounts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "operationId": "AccountsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "accounts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/accounts/{id}/metrics": {
      "get": {
        "operationId": "AccountsController_getAccountMetrics",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "accounts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/accounts/{id}/posts": {
      "get": {
        "operationId": "AccountsController_getAccountPosts",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Sayfalama için cursor (son gönderinin IDsi)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Sayfa başına kayıt sayısı",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "description": "Sıralama alanı",
            "schema": {
              "default": "date",
              "enum": [
                "date",
                "engagement"
              ],
              "type": "string"
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "description": "Sıralama yönü",
            "schema": {
              "default": "desc",
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string"
            }
          },
          {
            "name": "contentType",
            "required": false,
            "in": "query",
            "description": "İçerik tipi filtresi (örn. IMAGE, VIDEO, CAROUSEL)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "accounts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/accounts/{id}/overview": {
      "get": {
        "operationId": "AccountsController_getAccountOverview",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "range",
            "required": false,
            "in": "query",
            "description": "Analiz aralığı",
            "schema": {
              "default": "30d",
              "enum": [
                "7d",
                "30d",
                "90d"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "accounts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/accounts/{id}/sentiment": {
      "get": {
        "operationId": "AccountsController_getSentimentBreakdown",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "accounts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/accounts/{id}/hashtags": {
      "get": {
        "operationId": "AccountsController_getHashtagAnalysis",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "accounts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/accounts/{id}/topics": {
      "get": {
        "operationId": "AccountsController_getTopics",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "accounts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/accounts/{id}/best-times": {
      "get": {
        "operationId": "AccountsController_getBestTimes",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "accounts"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    }
  },
  "info": {
    "title": "Instascope API",
    "description": "Instascope backend servis dokümantasyonu",
    "version": "1.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "accounts",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Kullanıcının adı (isteğe bağlı)",
            "example": "Nazgül Aksoy"
          },
          "email": {
            "type": "string",
            "description": "Kullanıcının e-posta adresi",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "Kullanıcının şifresi",
            "example": "S3cur3P@ssw0rd"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Kullanıcının e-posta adresi",
            "example": "user@example.com"
          },
          "password": {
            "type": "string",
            "description": "Kullanıcının şifresi",
            "example": "S3cur3P@ssw0rd"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "description": "Yenileme tokenı",
            "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
          }
        },
        "required": [
          "refreshToken"
        ]
      },
      "CreateAccountDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "description": "Eklenecek Instagram kullanıcı adı",
            "example": "atolye.studio"
          },
          "sourceType": {
            "type": "string",
            "description": "Kaynak türü",
            "example": "API",
            "enum": [
              "API",
              "SCRAPE",
              "MOCK",
              "AI"
            ]
          },
          "frequency": {
            "type": "string",
            "description": "Toplama sıklığı",
            "example": "daily"
          },
          "accessTokenEnc": {
            "type": "string",
            "description": "İsteğe bağlı erişim anahtarı"
          }
        },
        "required": [
          "username",
          "sourceType"
        ]
      },
      "AccountResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Hesap IDsi",
            "example": "b7351c3c-c464..."
          },
          "igUsername": {
            "type": "string",
            "description": "Instagram kullanıcı adı",
            "example": "atolye.studio"
          },
          "igAccountId": {
            "type": "string",
            "description": "Instagram Hesap IDsi",
            "example": "17841444785369518"
          },
          "sourceType": {
            "type": "string",
            "description": "Kaynak türü",
            "enum": [
              "API",
              "SCRAPE",
              "MOCK",
              "AI"
            ]
          },
          "scheduleCron": {
            "type": "string",
            "description": "Zamanlama kuralı",
            "example": "0 0 * * *"
          },
          "status": {
            "type": "string",
            "description": "Hesap durumu",
            "example": "active"
          }
        },
        "required": [
          "id",
          "igUsername",
          "sourceType",
          "status"
        ]
      }
    }
  }
}